Reworked train controller GUI.
The new TrainControllerInstances class keeps a record of all existing
instances of the train controller UI. I changed TTEHomeGUI a bit to
reflect this change. I also created a TestPanel class but it's empty for
now because I figured that I need to commit and push this change as
early as possible.